Create Lab 3 Akshay.ipynb
Doubts in 4.c and 4.f
Also in 5.a

4.b can maybe made faster? But i don't know how

Overall, a very beautiful exercise. I still don't think I have fully grasped it, but very elegant and certainly thoughtful.

I will commit this in a parallel branch (not main), so that you don't overwrite your local code by mistake. Will raise a pull request for you to merge this with the Main branch instead